// src/generated-types/menu-category.ts
export interface MenuCategoryListWithItems {
  data: MenuCategoryWithItems[];
}

/** Message representing a list of menu categories with translations */
export interface MenuCategoryListWithTranslation {
  data: MenuCategoryWithTranslation[];
}

/** Message representing a menu category */
export interface MenuCategory {
  id: string;
  slug: string;
  position: number;
  imageUrl?: string | null | undefined;
  isAvailable: boolean;
  createdAt: Date | null;
  updatedAt: Date | null;
}

/** Message representing a menu category translation */
export interface MenuCategoryTranslation {
  id: string;
  language: string;
  title: string;
  description?: string | null | undefined;
}

/** Message representing a menu category with translation */
export interface MenuCategoryWithTranslation {
  id: string;
  slug: string;
  position: number;
  imageUrl?: string | null | undefined;
  isAvailable: boolean;
  translations: MenuCategoryTranslation[];
}

/** Message representing a menu category with items */
export interface MenuCategoryWithItems {
  id: string;
  slug: string;
  position: number;
  imageUrl?: string | null | undefined;
  isAvailable: boolean;
  translations: MenuCategoryTranslation[];
  menuItems: MenuItemWithTranslation[];
}

/** Request message for creating a menu category. */
export interface CreateMenuCategoryRequest {
  slug: string;
  imageUrl?: string | null | undefined;
  isAvailable?: boolean | null | undefined;
}

/** Request message for updating a menu category. */
export interface UpdateMenuCategoryRequest {
  id: string;
  slug?: string | null | undefined;
  imageUrl?: string | null | undefined;
  isAvailable?: boolean | null | undefined;
}

/** Request message for changing the position of a menu category. */
export interface ChangeMenuCategoryPositionRequest {
  id: string;
  position: number;
}

/** Request message for creating a menu category translation. */
export interface CreateMenuCategoryTranslationRequest {
  categoryId: string;
  language: string;
  title: string;
  description?: string | null | undefined;
}

/** Request message for updating a menu category translation. */
export interface UpdateMenuCategoryTranslationRequest {
  id: string;
  title?: string | null | undefined;
  description?: string | null | undefined;
}

/** Message representing a status response */
export interface StatusResponse {
  success: boolean;
  message: string;
}